Extract shared components and containers
