+++ app/index.html
@@ -3,6 +3,7 @@
   <head>
     <meta charset="UTF-8">
     <title>Classroom Desktop 0.0.1</title>
+    <link rel="stylesheet" href="../node_modules/octicons/build/octicons.css" charset="utf-8">
     <link rel="stylesheet" href="../node_modules/bootstrap/dist/css/bootstrap.min.css" charset="utf-8">
   </head>
