+++ app/routes/select-submissions/components/Submission.jsx
@@ -1,9 +1,5 @@
 import React, { PropTypes } from "react"
-
-const containerStyles = {
-  padding: "10px",
-  borderTop: "solid 1px #e3e3e3"
-}
+import ItemPanel from "../../../shared-components/ItemPanel"
 
 const Submission = ({
   id,
@@ -13,30 +9,14 @@ const Submission = ({
   repoUrl,
   selected
 }) => (
-  <div style={containerStyles}>
-    <div className="media">
-      <a className="media-left" href="#">
-        <img
-          width="48px"
-          height="48px"
-          src={avatarUrl}
-          className="media-object"
-          alt={username}/>
-      </a>
-      <div className="media-body">
-        <div className="row">
-          <div className="col-xs-6">
-            <h4 className="media-heading">@{username}</h4>
-            {displayName}
-          </div>
-          <div className="col-xs-6">
-            <a href={repoUrl}>{repoUrl}</a>
-            <input type="checkbox" checked={selected} readOnly={true}/>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
+  <ItemPanel
+    imagePath={avatarUrl}
+    title={displayName}
+    subtitle={username}
+  >
+    <a href={repoUrl}>{repoUrl}</a>
+    <input type="checkbox" checked={selected} readOnly={true}/>
+  </ItemPanel>
 )
 
 Submission.propTypes = {
