+++ app/shared-containers/AssignmentPanel.jsx
